+++ modules/mauth-signer-akka-http/src/example/resources/application.conf
@@ -1,0 +1,4 @@
+app {
+  uuid: ${?APP_MAUTH_UUID}
+  private_key: ${?APP_MAUTH_PRIVATE_KEY}
+}
