+++ modules/mauth-signer-apachehttp/README.md
@@ -6,7 +6,7 @@ This is an implementation of Medidata Authentication Client Signer to sign the H
 
 1. Configuration
    * MAuth uses [Typesafe Config](https://github.com/typesafehub/config).
-     Create `application.conf` on your classpath with following content.
+     Create `application.conf` on your classpath with the following content.
    
            app {
                uuid: "aaaa-bbbbb-ccccc-ddddd-eeeee"
@@ -20,9 +20,6 @@ This is an implementation of Medidata Authentication Client Signer to sign the H
                uuid: ${?APP_MAUTH_UUID}
                private_key: ${?APP_MAUTH_PRIVATE_KEY}
            }
-     * Load Configuration
-   
-           SignerConfiguration configuration = new SignerConfiguration(ConfigFactory.load());
 2. Signing Requests 
-   * To sign requests using Apache HttpClient interceptors, please see [MauthRequestInterceptorSignerExample](src/example/MauthRequestInterceptorSignerExample.java)
-   * To sign requests manually, please see [ManualSignerExample](src/example/ManualSignerExample.java)
+   * To sign requests using Apache HttpClient interceptors, please see [com.mdsol.mauth.MauthRequestInterceptorSignerExample](src/example/java/com/mdsol/mauth/MauthRequestInterceptorSignerExample.java)
+   * To sign requests manually, please see [com.mdsol.mauth.ManualSignerExample](src/example/java/com/mdsol/mauth/ManualSignerExample.java)
