+++ modules/mauth-signer-apachehttp/src/example/java/com/mdsol/mauth/ManualSignerExample.java
@@ -1,6 +1,6 @@
-import com.mdsol.mauth.SignerConfiguration;
+package com.mdsol.mauth;
+
 import com.mdsol.mauth.apache.HttpClientRequestSigner;
-import com.mdsol.mauth.apache.SignerHttpRequestInterceptor;
 import com.typesafe.config.ConfigFactory;
 import org.apache.http.StatusLine;
 import org.apache.http.client.methods.CloseableHttpResponse;
@@ -10,18 +10,15 @@ import org.apache.http.impl.client.HttpClients;
 
 import java.io.IOException;
 
-public class MauthRequestInterceptorSignerExample {
+public class ManualSignerExample {
 
-    public void executeMe() {
+    private void executeMe() {
         SignerConfiguration configuration = new SignerConfiguration(ConfigFactory.load());
         final HttpClientRequestSigner httpClientRequestSigner = new HttpClientRequestSigner(configuration);
-        final SignerHttpRequestInterceptor signerHttpRequestInterceptor = new SignerHttpRequestInterceptor(httpClientRequestSigner);
-        CloseableHttpClient httpClient = HttpClients
-                .custom()
-                .addInterceptorFirst(signerHttpRequestInterceptor)
-                .build();
-
         HttpGet request = new HttpGet("https://api.mdsol.com/v1/countries");
+        httpClientRequestSigner.signRequest(request);
+
+        CloseableHttpClient httpClient = HttpClients.custom().build();
 
         try(CloseableHttpResponse response = httpClient.execute(request)) {
             StatusLine status = response.getStatusLine();
@@ -33,7 +30,7 @@ public class MauthRequestInterceptorSignerExample {
     }
 
     /**
-     * Example how to use Medidata Authentication Client Signer to sign the Http requests
+     * Example how to sign requests manually
      * Set up the following environment variables:
      * APP_MAUTH_UUID - app uuid
      * APP_MAUTH_PRIVATE_KEY - the application private key itself, not the path
@@ -41,6 +38,6 @@ public class MauthRequestInterceptorSignerExample {
      * @param args - no args expected
      */
     public static void main(String[] args) {
-        new MauthRequestInterceptorSignerExample().executeMe();
+        new ManualSignerExample().executeMe();
     }
 }
