+++ project/Dependencies.scala
@@ -2,15 +2,13 @@ import sbt._
 
 object Dependencies {
 
-  def compile(deps: ModuleID*): Seq[ModuleID] = deps map (_ % "compile")
+  def compile(deps: ModuleID*): Seq[ModuleID] = deps map (_ % Compile)
 
-  def provided(deps: ModuleID*): Seq[ModuleID] = deps map (_ % "provided")
+  def provided(deps: ModuleID*): Seq[ModuleID] = deps map (_ % Provided)
 
-  def test(deps: ModuleID*): Seq[ModuleID] = deps map (_ % "test")
+  def test(deps: ModuleID*): Seq[ModuleID] = deps map (_ % Test)
 
-  def runtime(deps: ModuleID*): Seq[ModuleID] = deps map (_ % "runtime")
-
-  def container(deps: ModuleID*): Seq[ModuleID] = deps map (_ % "container")
+  def runtime(deps: ModuleID*): Seq[ModuleID] = deps map (_ % Runtime)
 
   private object Version {
     val akkaHttp = "10.0.10"
@@ -18,6 +16,9 @@ object Dependencies {
   }
 
   val akkaHttp: ModuleID = "com.typesafe.akka" %% "akka-http" % Version.akkaHttp
+  val akkaActor: ModuleID = "com.typesafe.akka" %% "akka-actor" % "2.4.19"
+  val akkaStream: ModuleID = "com.typesafe.akka" %% "akka-stream" % "2.4.19"
+  val akkaHttpCore: ModuleID = "com.typesafe.akka" %% "akka-http-core" % Version.akkaHttp
   val apacheHttpClient: ModuleID = "org.apache.httpcomponents" % "httpclient" % "4.5.5"
   val bouncyCastlePkix: ModuleID = "org.bouncycastle" % "bcpkix-jdk15on" % "1.59"
   val commonsCodec: ModuleID = "commons-codec" % "commons-codec" % "1.11"
